refactor: move Search component
